# pkg/api/api.yml
openapi: "3.0.3"
info:
  version: "1.0"
  title: "Cominify API"
  description: API for Cominify chatbot
servers:
  - url: http://localhost:8080/api/v1
    description: Local Test Server
paths:
  /users:
    get:
      description: "Return list of users"
      operationId: Users
      tags:
        - user
      responses:
        "200":
          description: Successfully return a list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      description: "Return list of users"
      operationId: GetUser
      tags:
        - user
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successfully return a list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}/comics:
    get:
      description: "Return list of comics which user subscribed to"
      operationId: GetUserComics
      tags:
        - comic
      parameters:
        - name: id
          in: path
          description: User App ID, different with User Page Scope ID
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/q"
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Successfully return a list of comics which user subscribed to
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comic"
        "404":
          description: User not found
    # post:
    #   operationId: SubscribeComic
    #   description: User subscribe to new comic
    #   tags:
    #     - comic
    #   parameters:
    #     - name: id
    #       in: path
    #       description: User App ID, different with User Page Scope ID
    #       required: true
    #       schema:
    #         type: string
    #   requestBody:
    #     description: "Comic link"
    #     content:
    #       text/plain:
    #         schema:
    #           type: string
    #           example: https://beeng.net/truyen-tranh-online/toi-thang-cap-mot-minh-ss2-33790
    #   responses:
    #     "200":
    #       description: Successfully subscribed to a comic
    #       content:
    #         application/json:
    #           schema:
    #             $ref: "#/components/schemas/Comic"
    #     "404":
    #       description: User not found
    #     "401":
    #       description: User unauthorized
    #     "500":
    #       description: Internal error
  /users/{user_id}/comics/{id}:
    delete:
      description: Unsubscribe comic
      operationId: UnsubscribeComic
      tags:
        - comic
      parameters:
        - name: user_id
          in: path
          description: User App ID, different with User Page Scope ID
          required: true
          schema:
            type: string
        - name: id
          in: path
          description: Comic ID
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unsubscribe comic responses
        "404":
          description: Comic not found
  /comics:
    get:
      # summary: "Get list of comics"
      description: "Return list of comics"
      operationId: Comics
      tags:
        - comic
      parameters:
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Successfully return a list of comics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComicPage"
  /comics/{id}:
    get:
      description: "Return comics based on ID"
      operationId: GetComic
      tags:
        - comic
      parameters:
        - name: id
          in: path
          description: Comic ID
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Found comic with ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comic"
        "404":
          description: Comics does not exists
components:
  parameters:
    q:
      name: q
      in: query
      description: Used to query by name in a list operation.
      schema:
        type: string
        default: ""
    offset:
      name: offset
      in: query
      description: Used to request the next page in a list operation.
      schema:
        type: integer
        format: int64
        default: 0
    limit:
      name: limit
      in: query
      description: Used to specify the maximum number of records which are returned in the next page.
      schema:
        type: integer
        format: int64
        default: 50
  schemas:
    Comic:
      type: object
      properties:
        id:
          type: integer
          description: Comic ID
        page:
          type: string
          description: Page to read comic
        name:
          type: string
          description: Comic name
        url:
          type: string
          description: Comic url
        latestChap:
          type: string
          description: Comic's latest chapter
        imgURL:
          type: string
          description: URL to comic's avatar
        chapURL:
          type: string
          description: Chapter url
    ComicPage:
      description: List comic response
      required:
        - comics
      properties:
        comics:
          type: array
          items:
            $ref: "#/components/schemas/Comic"
    User:
      type: object
      properties:
        name:
          type: string
          description: Comic name
        psid:
          type: string
          description: Page scope ID
        appid:
          type: string
          description: App scope ID
        profile_pic:
          type: string
          description: User avatar link
        comics:
          type: integer
          description: Number of comics subscribed