# pkg/db/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.

package db

import (
	"database/sql"
	"time"
)

type Comic struct {
	ID          int32
	Page        string
	Name        string
	Url         string
	ImgUrl      string
	CloudImgUrl string
	LatestChap  string
	ChapUrl     string
}

type Subscriber struct {
	ID        int32
	UserPsid  string
	UserAppid string
	ComicID   int32
	CreatedAt time.Time
}

type User struct {
	ID         int32
	Name       string
	Psid       sql.NullString
	Appid      sql.NullString
	ProfilePic sql.NullString
}